refactor(migrate): modularize template_compiler
Split 2,275-line template_compiler/mod.rs into 9 focused files:
- compilation.rs (514 lines): occurrence collection and merging
- deduplication.rs (499 lines): dedup, variable tracking
- node_compiler.rs (330 lines): CslnNode → TemplateComponent
- formatting.rs (326 lines): rendering, wrapping, delimiters
- bibliography.rs (171 lines): bibliography and type templates
- types.rs (161 lines): ItemType collection and mapping
- overrides.rs (103 lines): type-specific override logic
- sorting.rs (74 lines): citation/bibliography ordering
- mod.rs (73 lines): public API and shared types

Public API unchanged. All 404 tests pass. No logic changes.
